Refactor functional
